hybridswap: avoid unsigned underflow in zram watermark min()
zram_watermark_ok() computes wm as min(fetch_zram_wm_scale_value(),
fetch_zram_wm_scale_value() - diff_buffers). The first operand is u64
while the second is long long, so diff_buffers is promoted to unsigned.
When diff_buffers exceeds the scale value, the subtraction wraps to a
huge positive number, min() always selects the scale value, and
zram_watermark_ok() wrongly returns false under real memory pressure,
suppressing swapout of the zram device.

Cast both operands to long long so a negative wm correctly forces
zram_watermark_ok() to true and swapd acts on the pressure.

+++ drivers/moto_swap/hybridswap/hybridswap_swapd.c
@@ -1262,7 +1262,8 @@ bool zram_watermark_ok(void)
 	diff_buffers = diff_buffers * percent_constant / fetch_nr_zram_total();
 
 	cur_scale = zram_used * percent_constant / fetch_nr_zram_total();
-	wm  = min(fetch_zram_wm_scale_value(), fetch_zram_wm_scale_value()- diff_buffers);
+	wm  = min((long long)fetch_zram_wm_scale_value(),
+		  (long long)fetch_zram_wm_scale_value() - diff_buffers);
 
 	return cur_scale > wm;
 }
